+++ builders/StrBuilder.d.ts
@@ -5,4 +5,5 @@ export default class StrBuilder {
     toLowerCase(combine?: boolean): StrBuilder | string;
     toPascalCase(combine?: boolean): StrBuilder | string;
     toUpperCase(combine?: boolean): StrBuilder | string;
+    random(size?: number): string;
 }