+++ facades/Str.d.ts
@@ -3,4 +3,5 @@ export default class Str {
     static toLowerCase(value: string, combine?: boolean): StrBuilder | string;
     static toPascalCase(value: string, combine?: boolean): StrBuilder | string;
     static toUpperCase(value: string, combine?: boolean): StrBuilder | string;
+    static random(size?: number): string;
 }