+++ src/facades/Str.ts
@@ -12,4 +12,8 @@ export default class Str {
     public static toUpperCase(value: string, combine?: boolean): StrBuilder | string {
         return new StrBuilder().setValue(value).toUpperCase(combine);
     }
+
+    public static random(size?: number): string {
+        return new StrBuilder().random(size);
+    }
 }